requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
python-dotenv==1.0.0
typing_extensions==4.9.0
yarl==1.9.4
